Moved add_*_args to `__main__.py`
