+++ tests/test_init.py
@@ -9,13 +9,6 @@ import pytest
 
 TEST_RESULTS_PATH = "test_results"
 
-class TestInitGitCheck:
-    def test_fail_if_not_in_git_repo(self, tmp_path, monkeypatch) -> None:
-        """check_git() should raise an error if the current directory is not a git repository"""
-        monkeypatch.chdir(tmp_path) # No git project in tmp_path
-
-        with pytest.raises(RuntimeError, match="The current project is not a git repository. Try running 'git init' or clone a GitHub project."):
-            check_git()
 
 class TestInitCommand:
     def test_fail_if_not_under_home(self, tmp_path, monkeypatch) -> None:
@@ -25,6 +18,7 @@ class TestInitCommand:
 
         with pytest.raises(RuntimeError, match="cluv init should be run in a directory under your home directory."):
             init()
+
 
     def test_generate_default_toml_config(self, tmp_path, monkeypatch) -> None:
         """init() should create a pyproject.toml file with the default configuration if it doesn't exist"""
@@ -38,6 +32,7 @@ class TestInitCommand:
         assert config.results_path == "logs"
         assert config.slurm == {'UV_OFFLINE': 1, 'WANDB_MODE': 'offline'}
         assert config.cluster_configs == {"mila": {"UV_OFFLINE": 0, "WANDB_MODE": "online"}}
+
 
     def test_keep_toml_config(self, tmp_path, monkeypatch) -> None:
         """init() should keep the cluv config of an already existing pyproject.toml"""
@@ -55,3 +50,13 @@ results_path = "results"
 
         assert config.results_path == "results"
         assert config.clusters == ["mila"]
+
+
+class TestInitGitCheck:
+    def test_fail_if_not_in_git_repo(self, tmp_path, monkeypatch) -> None:
+        """check_git() should raise an error if the current directory is not a git repository"""
+        monkeypatch.chdir(tmp_path) # No git project in tmp_path
+
+        with pytest.raises(RuntimeError, match="The current project is not a git repository. Try running 'git init' or clone a GitHub project."):
+            check_git()
+
